+++ src/main/java/org/opentmf/mockserver/callback/DynamicJsonPatchCollectionCallback.java
@@ -55,8 +55,7 @@ public class DynamicJsonPatchCollectionCallback implements ExpectationResponseCa
 
   @Override
   public HttpResponse handle(HttpRequest httpRequest) {
-    HttpResponse authError =
-        TokenEnforcer.getInstance().validateWithRoles(httpRequest, "writer", "admin");
+    HttpResponse authError = TokenEnforcer.getInstance().validateForRequest(httpRequest);
     if (authError != null) {
       return authError;
     }