+++ src/main/java/org/opentmf/mockserver/callback/DynamicPostCallback.java
@@ -103,8 +103,7 @@ public class DynamicPostCallback implements ExpectationResponseCallback {
 
   @Override
   public HttpResponse handle(HttpRequest httpRequest) {
-    HttpResponse authError =
-        TokenEnforcer.getInstance().validateWithRoles(httpRequest, "writer", "admin");
+    HttpResponse authError = TokenEnforcer.getInstance().validateForRequest(httpRequest);
     if (authError != null) {
       return authError;
     }